Add selected text example, rename whole page example class
